// front-End/src/app/model/ideal-learn-obj.ts
export class IdealLearnObj {
    id: number;
    instance_name: string;
    concept: string;
    title: string;
    description: string;
    interactivity_type: string;
    learn_resource_type: string;
    interactivity_level: string;
    semantic_density: string;
    difficulty: string;
    quality: number;
}